refactor: some refactoring
