+++ src/app/api/moviesApi.ts
@@ -1,4 +1,4 @@
-import { SortType, getReviewsType } from '../../utils/types';
+import { SortType } from '../../utils/types';
 import { baseApi } from './baseApi';
 
 type getMoviesType = {
@@ -27,10 +27,10 @@ export const moviesApi = baseApi.injectEndpoints({
       query: id => `movie/${id}`,
     }),
     getReviews: build.query({
-      query: ({ id }: getReviewsType) => `review?page=1&limit=10&id=${id}`,
+      query: id => `review?page=1&limit=10&id=${id}`,
     }),
     getPosters: build.query({
-      query: ({ id }) => `image?page=1&limit=10&movieId=${id}`,
+      query: id => `image?page=1&limit=10&movieId=${id}`,
     }),
   }),
 });
